+++ ui/src/components/os/index.ts
@@ -2,6 +2,7 @@ export { default as Header } from './Header'
 export { default as Sidebar } from './Sidebar'
 export { default as StatusIndicator } from './StatusIndicator'
 export { default as ApplicationPanel } from './ApplicationPanel'
+export { default as ConsolePanel } from './ConsolePanel'
 export { default as KernelStatusPanel } from './KernelStatusPanel'
 export { default as FilesystemPanel } from './FilesystemPanel'
 export { default as LogsPanel } from './LogsPanel'
@@ -11,6 +12,7 @@ export type {
   KernelStats,
   FilesystemStats,
   DirEntry,
+  ConsoleLine,
   PanelType,
   StatusType,
   LogEntry,